feat: Setup project structure for refactor of repositories
