+++ MiniTwit.Core/IMongoDBContext.cs
@@ -1,0 +1,1 @@
+namespace MiniTwit.Core;
