+++ MiniTwit.Core/IRepositories/ITweetRepository.cs
@@ -1,0 +1,1 @@
+namespace MiniTwit.Core.IRepositories;
