+++ MiniTwit.Core/IServices/IMongoDBService.cs
@@ -1,4 +1,4 @@
-namespace MiniTwit.Server.Service;
+namespace MiniTwit.Core.IServices;
 
 public interface IMongoDBService
 {
