+++ MiniTwit.Infrastructure/MongoDBContext.cs
@@ -1,0 +1,1 @@
+namespace MiniTwit.Infrastructure;
