+++ MiniTwit.Service/Services/MongoDBService.cs
@@ -1,0 +1,1 @@
+namespace MiniTwit.Service.Services;
