+++ itu-minitwit.sln
@@ -7,6 +7,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MiniTwit.Server", "MiniTwit
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MiniTwit.Server.Tests", "MiniTwit.Server.Tests\MiniTwit.Server.Tests.csproj", "{7CF7D7F0-1603-4550-93BA-2CD68839DB65}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MiniTwit.Core", "MiniTwit.Core\MiniTwit.Core.csproj", "{7D5F6EF5-A157-4D1C-82B7-6C8405003605}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MiniTwit.Infrastructure", "MiniTwit.Infrastructure\MiniTwit.Infrastructure.csproj", "{8DBEB98F-3258-4FFB-AD7D-ACC2327BBE6F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MiniTwit.Service", "MiniTwit.Service\MiniTwit.Service.csproj", "{0653CE2D-A13A-4C87-9DB8-35D126C607D4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -24,5 +30,17 @@ Global
 		{7CF7D7F0-1603-4550-93BA-2CD68839DB65}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7CF7D7F0-1603-4550-93BA-2CD68839DB65}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7CF7D7F0-1603-4550-93BA-2CD68839DB65}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7D5F6EF5-A157-4D1C-82B7-6C8405003605}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7D5F6EF5-A157-4D1C-82B7-6C8405003605}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7D5F6EF5-A157-4D1C-82B7-6C8405003605}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7D5F6EF5-A157-4D1C-82B7-6C8405003605}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8DBEB98F-3258-4FFB-AD7D-ACC2327BBE6F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8DBEB98F-3258-4FFB-AD7D-ACC2327BBE6F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8DBEB98F-3258-4FFB-AD7D-ACC2327BBE6F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8DBEB98F-3258-4FFB-AD7D-ACC2327BBE6F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0653CE2D-A13A-4C87-9DB8-35D126C607D4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0653CE2D-A13A-4C87-9DB8-35D126C607D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0653CE2D-A13A-4C87-9DB8-35D126C607D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0653CE2D-A13A-4C87-9DB8-35D126C607D4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
